+++ app/src/main/java/com/example/asus/example/mvvm/ViewModel/GroupViewModel.java
@@ -33,7 +33,7 @@ public class GroupViewModel extends ViewModel {
         UserRepository userRepository = new UserRepository();
         groupRepository = new GroupRepository();
         SharedPreferences myPrefs = context.getSharedPreferences("CurrentUser", 0);
-        currentUser = userRepository.getUserByID(myPrefs.getLong("CurrentUserId", 0)).getValue();
+        currentUser = userRepository.getUserByID(myPrefs.getInt("CurrentUserId", 0)).getValue();
     }
 
     public MutableLiveData<List<Group>> getGroups() {
