+++ app/src/main/java/com/example/asus/example/mvvm/ViewModel/ItemGroupViewModel.java
@@ -46,7 +46,7 @@ public class ItemGroupViewModel extends ViewModel {
         groupRepository = new GroupRepository();
 
         SharedPreferences myPrefs = context.getSharedPreferences("CurrentUser", 0);
-        currentUser = userRepository.getUserByID(myPrefs.getLong("CurrentUserId", 0)).getValue();
+        currentUser = userRepository.getUserByID(myPrefs.getInt("CurrentUserId", 0)).getValue();
     }
