+++ app/src/main/java/com/example/asus/example/mvvm/ViewModel/ItemPostViewModel.java
@@ -38,7 +38,7 @@ public class ItemPostViewModel extends ViewModel {
         UserRepository userRepository = new UserRepository();
 
         SharedPreferences myPrefs = context.getSharedPreferences("CurrentUser", 0);
-        currentUser = userRepository.getUserByID(myPrefs.getLong("CurrentUserId", 0)).getValue();
+        currentUser = userRepository.getUserByID(myPrefs.getInt("CurrentUserId", 0)).getValue();
     }
