+++ app/src/main/java/com/example/asus/example/mvvm/ViewModel/ItemUserViewModel.java
@@ -44,7 +44,7 @@ public class ItemUserViewModel extends ViewModel {
         postRepository = new PostRepository();
 
         SharedPreferences myPrefs = context.getSharedPreferences("CurrentUser", 0);
-        currentUser = userRepository.getUserByID(myPrefs.getLong("CurrentUserId", 0)).getValue();
+        currentUser = userRepository.getUserByID(myPrefs.getInt("CurrentUserId", 0)).getValue();
     }
 
     @BindingAdapter({"currentUserImageUrl"})
