+++ app/src/main/java/com/example/asus/example/mvvm/ViewModel/UserViewModel.java
@@ -34,7 +34,7 @@ public class UserViewModel extends ViewModel {
         userRepository = new UserRepository();
 
         SharedPreferences myPrefs = context.getSharedPreferences("CurrentUser", 0);
-        currentUser = userRepository.getUserByID(myPrefs.getLong("CurrentUserId", 0)).getValue();
+        currentUser = userRepository.getUserByID(myPrefs.getInt("CurrentUserId", 0)).getValue();
     }
 
     /**
